// UT4/ExUT4_0_Angular/src/app/interfaces/fetch-data.ts
/**
 * Interfaz para recoger los datos de la API
 */
export interface FetchData {
  results: [
    {
      name: string;
      rotation_period: string;
      orbital_period: string;
      diameter: string;
      climate: string;
      gravity: string;
      terrain: string;
      surface_water: string;
      population: string;
      residents: string[];
      films: string[];
      created: string;
      edited: string;
      url: string;
    }
  ]
}